# docs/openapi/v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Adwais.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/monitors/analytics": {
      "get": {
        "tags": [
          "Monitor"
        ],
        "summary": "Unified analytics endpoint for monitoring data.\r\nProvides latency time-series and filtered monitor lists hydrated with uptime for the specified timeframe (defaults to T30).",
        "parameters": [
          {
            "name": "tenantId",
            "in": "query",
            "description": "Optional. Scopes the metrics to a specific tenant. \r\nIf null, metrics represent the global portfolio total.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "monitorId",
            "in": "query",
            "description": "Optional. Scopes the metrics to a specific monitor. \r\nIf null, metrics represent the global portfolio total.\r\nOverrides tenantId parameter.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "description": "The primary timeframe for the calculation (e.g., T7, T30). \r\nDefaults to T30.",
            "schema": {
              "$ref": "#/components/schemas/Timeframe"
            }
          },
          {
            "name": "comparison",
            "in": "query",
            "description": "The comparison period for the timeframe.\r\nDefaults to Preceding.",
            "schema": {
              "$ref": "#/components/schemas/ComparisonType"
            }
          },
          {
            "name": "tags",
            "in": "query",
            "description": "Optional. Filters the monitors by specific tags.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "statuses",
            "in": "query",
            "description": "Optional. Filters the monitors by specific statuses.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorAnalyticsResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorAnalyticsResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorAnalyticsResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/monitors": {
      "get": {
        "tags": [
          "Monitor"
        ],
        "summary": "Retrieves monitors, optionally filtered by tenant or specific monitor ID.\r\nReturns monitors hydrated with uptime for the specified timeframe (defaults to T30).",
        "parameters": [
          {
            "name": "tenantId",
            "in": "query",
            "description": "Optional. Scopes the metrics to a specific tenant. \r\nIf null, metrics represent the global portfolio total.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "monitorId",
            "in": "query",
            "description": "Optional. Scopes the metrics to a specific monitor. \r\nIf null, metrics represent the global portfolio total.\r\nOverrides tenantId parameter.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "description": "The primary timeframe for the calculation (e.g., T7, T30). \r\nDefaults to T30.",
            "schema": {
              "$ref": "#/components/schemas/Timeframe"
            }
          },
          {
            "name": "comparison",
            "in": "query",
            "description": "The comparison period for the timeframe.\r\nDefaults to Preceding.",
            "schema": {
              "$ref": "#/components/schemas/ComparisonType"
            }
          },
          {
            "name": "tags",
            "in": "query",
            "description": "Optional. Filters the monitors by specific tags.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "statuses",
            "in": "query",
            "description": "Optional. Filters the monitors by specific statuses.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UptimeMonitorDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UptimeMonitorDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UptimeMonitorDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Monitor"
        ],
        "summary": "Creates a new uptime monitor in UptimeRobot and registers it in the system.",
        "parameters": [
          {
            "name": "tenantId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMonitorRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMonitorRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMonitorRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UptimeMonitorDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UptimeMonitorDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UptimeMonitorDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/monitors/unassigned": {
      "get": {
        "tags": [
          "Monitor"
        ],
        "summary": "Retrieves monitors that are not assigned to any specific tenant.",
        "parameters": [
          {
            "name": "timeframe",
            "in": "query",
            "description": "The timeframe for calculating uptime percentage.",
            "schema": {
              "$ref": "#/components/schemas/Timeframe"
            }
          },
          {
            "name": "comparison",
            "in": "query",
            "description": "The comparison period type.",
            "schema": {
              "$ref": "#/components/schemas/ComparisonType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UptimeMonitorDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UptimeMonitorDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UptimeMonitorDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/monitors/{id}/assign/{tenantId}": {
      "patch": {
        "tags": [
          "Monitor"
        ],
        "summary": "Reassigns a monitor to a different tenant.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/monitors/{id}/unassign": {
      "patch": {
        "tags": [
          "Monitor"
        ],
        "summary": "Moves a monitor to the unassigned (system) tenant.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/monitors/{id}/pause": {
      "post": {
        "tags": [
          "Monitor"
        ],
        "summary": "Pauses monitoring for a specific monitor in UptimeRobot.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/monitors/{id}/start": {
      "post": {
        "tags": [
          "Monitor"
        ],
        "summary": "Resumes monitoring for a specific monitor in UptimeRobot.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/monitors/{id}": {
      "delete": {
        "tags": [
          "Monitor"
        ],
        "summary": "Deletes a monitor from both the system and UptimeRobot.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "tenantId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Monitor"
        ],
        "summary": "Updates monitor properties, such as SLA.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMonitorRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMonitorRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMonitorRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UptimeMonitorDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UptimeMonitorDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UptimeMonitorDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/monitors/{id}/latency": {
      "get": {
        "tags": [
          "Monitor"
        ],
        "summary": "Retrieves aggregated latency metrics for a specific monitor.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "tenantId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LatencyMetricsDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LatencyMetricsDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LatencyMetricsDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComparisonType": {
        "enum": [
          "Preceding",
          "YearOverYear"
        ],
        "type": "string"
      },
...
      "CreateMonitorRequestDto": {
        "required": [
          "name",
          "url"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string",
            "description": "The display name for the monitor."
          },
          "url": {
            "minLength": 1,
            "type": "string",
            "description": "The absolute URL or host to monitor."
          },
          "type": {
            "type": "string",
            "description": "Optional UptimeRobot monitor type. Defaults to HTTP when omitted.",
            "nullable": true
          },
          "uptimeSla": {
            "type": "number",
            "description": "Optional target uptime percentage (0-100).",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Data transfer object for creating a new uptime monitor."
      },
...
      "LatencyMetricsDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "average": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "lowest": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "highest": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LatencyPointResponseDto": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "average": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "previousAverage": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "p10": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "p90": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "MonitorAnalyticsResponseDto": {
        "type": "object",
        "properties": {
          "globalAverageLatency": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "latencyPoints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LatencyPointResponseDto"
            },
            "nullable": true
          },
          "monitors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UptimeMonitorDto"
            },
            "nullable": true
          },
          "kpis": {
            "$ref": "#/components/schemas/MonitorKpiResponseDto"
          }
        },
        "additionalProperties": false
      },
...
      "Timeframe": {
        "enum": [
          "Today",
          "T7",
          "T30",
          "T90",
          "Ytd",
          "T365"
        ],
        "type": "string"
      },
...
      "UpdateMonitorRequestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Friendly name for the monitor.",
            "nullable": true
          },
          "url": {
            "type": "string",
            "description": "URL for the monitor.",
            "nullable": true
          },
          "type": {
            "type": "string",
            "description": "UptimeRobot monitor type. Omit to leave the type unchanged.",
            "nullable": true
          },
          "sla": {
            "type": "number",
            "description": "SLA for the monitor.",
            "format": "double",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Tags associated with the monitor.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Request DTO updating a monitor."
      },
...
      "UptimeMonitorDto": {
        "required": [
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "tenantName": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "minLength": 1,
            "type": "string"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "updateInterval": {
            "type": "integer",
            "format": "int32"
          },
          "latencyDegradedFloor": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "uptimeSla": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "currentUptimePercentage": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "currentLatency": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "uptimeMonitorEnabled": {
            "type": "boolean"
          },
          "currentStatus": {
            "type": "string",
            "nullable": true
          },
          "lastUpdate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastUptimeUpdate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastLatencyUpdate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastSyncError": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Enter 'Bearer' [space] and then your valid token in the text input below.\r\n\r\nExample: \"Bearer eyJhbGciOiJIUzI1Ni...\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}